n55: test: Add test case when user don't have access to update or delete a record from a board they is not a member of

+++ e2e/finances/records/deleting.test.ts
@@ -2,12 +2,9 @@ import { IFinanceRecord } from "../../../src/interfaces/finance"
 import { authorize } from "../../helpers/authorize"
 import { fetchApi } from "../../helpers/fetchApi"
 
-beforeEach(async () => {
-  await authorize("jessica-stark")
-})
-
 describe("Finance record deleting", () => {
   it("returns a correct response after deleting", async () => {
+    await authorize("jessica-stark")
     const recordDeletingResponse = await fetchApi("/api/finances/records/5", { method: "DELETE" })
     expect(recordDeletingResponse.status).toEqual(200)
     expect(await recordDeletingResponse.json()).toEqual<IFinanceRecord>({
@@ -25,6 +22,7 @@ describe("Finance record deleting", () => {
   })
 
   it("the deleted records are not presented in all records list", async () => {
+    await authorize("jessica-stark")
     await fetchApi("/api/finances/records/1", { method: "DELETE" })
     await fetchApi("/api/finances/records/2", { method: "DELETE" })
     await fetchApi("/api/finances/records/3", { method: "DELETE" })
@@ -56,5 +54,12 @@ describe("Finance record deleting", () => {
         isTrashed: false,
       },
     ])
+  })
+
+  test("the user cannot delete a record of a board that they is not a member of", async () => {
+    await authorize("john-doe")
+    const recordUpdatingResponse = await fetchApi("/api/finances/records/5", { method: "DELETE" })
+    expect(recordUpdatingResponse.status).toEqual(403)
+    expect(await recordUpdatingResponse.json()).toEqual({ message: "Access denied." })
   })
 })
